refactor(scripts): extract json write responsibility from craftable recipe converter
