+++ scripts/recipe-json-converter/index.ts
@@ -9,9 +9,18 @@ import behavioursSchema from "./schemas/block-behaviours.json";
 import { JSONFileReader } from "./interfaces/json-file-reader";
 import { BlockBehaviours, Recipes, PiplizToolsets } from "./types";
 import { factory as createJSONFileAdapter } from "./adapters/json-file-adapter";
-import { convertRecipes } from "./domain/recipe-converter";
+import { convertRecipes as baseConvertCraftableRecipes } from "./domain/craftable-recipe-converter";
+import { convertRecipes as baseConvertRecipes } from "./domain/recipe-converter";
 import { findFiles } from "./adapters/fs-file-adapter";
 import { writeJSONToFile } from "./adapters/json-file-writer";
+import {
+    CraftableRecipeConverter,
+    CraftableRecipeConverterInputs,
+} from "./interfaces/craftable-recipe-converter";
+import {
+    RecipeConverter,
+    RecipeConverterInputs,
+} from "./interfaces/recipe-converter";
 
 const parser = yargs(hideBin(process.argv)).options({
     inputDirectory: {
@@ -42,21 +51,44 @@ const createRecipesFileReader = (): JSONFileReader<Recipes> => {
     return createJSONFileAdapter(recipesSchema);
 };
 
-(async () => {
-    const argv = await parser.argv;
+const createCraftableRecipeConverter = (): ((
+    input: CraftableRecipeConverterInputs
+) => ReturnType<CraftableRecipeConverter>) => {
     const toolsReader = createToolsFileReader();
     const behavioursReader = createBehavioursFileReader();
     const recipesReader = createRecipesFileReader();
+
+    return (input) =>
+        baseConvertCraftableRecipes({
+            ...input,
+            findFiles,
+            readToolsFile: toolsReader,
+            readBehavioursFile: behavioursReader,
+            readRecipeFile: recipesReader,
+        });
+};
+
+const createRecipeConverter = (): ((
+    input: RecipeConverterInputs
+) => ReturnType<RecipeConverter>) => {
+    const convertCraftableRecipes = createCraftableRecipeConverter();
+
+    return (input) =>
+        baseConvertRecipes({
+            ...input,
+            convertCraftableRecipes,
+            writeJSON: writeJSONToFile,
+        });
+};
+
+(async () => {
+    const argv = await parser.argv;
+    const convertRecipes = createRecipeConverter();
 
     try {
         const result = await convertRecipes({
             inputDirectoryPath: argv.inputDirectory,
             outputFilePath: argv.outputFile,
-            readToolsFile: toolsReader,
-            readBehavioursFile: behavioursReader,
-            readRecipeFile: recipesReader,
-            findFiles,
-            writeJSON: writeJSONToFile,
         });
 
         process.exit(result ? 0 : 1);
